Using diamond square algorithm to generate terrain
Works well for square maps, but is skewed when rectangular. I would like
to find a way to make rectangle maps not look skewed. I think by
actually generating a square map with sides equal to the largest
dimension of the rectangle map, then only showing what's needed, is the
way to go.